requirements.txt
pymupdf==1.26.7
openai==2.20.0
pydantic==2.5.3
flask==3.0.0
cryptography==41.0.7
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.3.2
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.67.3
